requirements.txt
beautifulsoup4>=4.11.1
bs4>=0.0.1
certifi>=2022.6.15
charset-normalizer>=2.1.1
click>=8.1.3
Flask>=2.2.2
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
idna>=3.3
itsdangerous>=2.1.2
Jinja2>=3.1.2
lxml>=4.9.1
MarkupSafe>=2.1.1
mypy-extensions>=0.4.3
pathspec>=0.9.0
platformdirs>=2.5.2
psycopg2-binary>=2.9.5
pydash>=5.1.0
requests>=2.28.1
rotten-tomatoes-scraper>=1.4.0
soupsieve>=2.3.2
tomli>=2.0.1
urllib3>=1.26.12
waitress>=2.1.2
Werkzeug>=2.2.2
